// types.ts
export enum CalculatorType {
  DRIP_RATE = 'Drip Rate',
  DOSING = 'Dosing',
  BSA = 'Body Surface Area',
  BODY_WEIGHT = 'Ideal & Adjusted Body Weight',
}

export enum Unit {
  KG = 'kg',
  LBS = 'lbs',
  CM = 'cm',
  IN = 'in',
  HR = 'hr',
  MIN = 'min',
}

export enum Sex {
    Male = 'Male',
    Female = 'Female'
}